+++ packages/utils/src/types/index.ts
@@ -1,0 +1,2 @@
+// SHOPLFLOW_DOCS_IGNORE
+export * from './$Values';